Adding hidden entities feature
Add a feature that allows a user to hide specific entities from the
default list of an entity type. A toggle switch is added to the entity
page, and controls whether the entity id is added or removed from the
list of hidden entities (kept locally). In addition, a new entities
view is added that contains all hidden entities.